may 15 events

+++ database/migrations/2015_11_07_203459_create_events_table.php
@@ -19,7 +19,7 @@ class CreateEventsTable extends Migration
             $table->string('name');
             $table->dateTime('startDate');
             $table->dateTime('endDate');
-            $table->integer('length_hours');
+            $table->decimal('length_hours');
             $table->integer('length_days');
             $table->boolean('allDay');
             $table->boolean('repeats');
